* merged changes from FreeBSD 9.0 for most wlan drivers except atheroswifi and net80211.
* added iwlwifi-6000g2b-ucode-18.168.6.1

+++ src/libs/compat/freebsd_wlan/net80211/ieee80211_ioctl.h
@@ -582,7 +582,7 @@ struct ieee80211req {
 	char		i_name[IFNAMSIZ];	/* if_name, e.g. "wi0" */
 	uint16_t	i_type;			/* req type */
 	int16_t		i_val;			/* Index or simple value */
-	int16_t		i_len;			/* Index or simple value */
+	uint16_t	i_len;			/* Index or simple value */
 	void		*i_data;		/* Extra data */
 };
 #define	SIOCS80211		 _IOW('i', 234, struct ieee80211req)
@@ -798,7 +798,7 @@ struct ieee80211req_scan_result {
 #endif
 	int8_t						isr_noise;
 	int8_t						isr_rssi;
-	uint8_t						isr_intval;		/* beacon interval */
+	uint16_t					isr_intval;		/* beacon interval */
 	uint8_t						isr_capinfo;		/* capabilities */
 	uint8_t						isr_erp;		/* ERP element */
 	uint8_t						isr_bssid[IEEE80211_ADDR_LEN];
